backend/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-multipart==0.0.9
Pillow==10.4.0
numpy==1.26.4
pandas==2.2.3
matplotlib==3.9.2
scikit-learn==1.5.2
xgboost==2.1.1
ultralytics==8.3.34
opencv-python-headless==4.10.0.84
open3d==0.18.0
trimesh==4.4.9
scikit-image==0.24.0
